refactor : userInput에서 InputValidator 클래스 분리
