+++ src/main/java/gmbs/UserInput.java
@@ -1,36 +1,16 @@
 package gmbs;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Scanner;
 
 public class UserInput {
     private static final String REPLAY_VALUE = "1";
     private static final String QUIT_VALUE = "2";
     private final Scanner scan = new Scanner(System.in);
+    private final InputValidator validator = new InputValidator();
 
 
-    public boolean isNaturalNumber(String input) {
-        try {
-            if (Integer.parseInt(input) > 0) {
-                return true;
-            }
-        } catch (Exception e) {
-            return false;
-        }
-        return false;
-    }
-
-    public boolean isValidLength(String input, int length) {
-        return input.length() == length;
-    }
-
-    public boolean hasNoOverlap(String input) {
-        Set<String> tempSet = new HashSet<>(Arrays.asList(input.split("")));
-        return tempSet.size() == input.length();
-    }
-
-    public boolean isValidNumberInput(String input, int length) {
-        return isNaturalNumber(input) && isValidLength(input, length) && hasNoOverlap(input);
-    }
 
     private String getInput() {
         return scan.nextLine();
@@ -39,7 +19,7 @@ public class UserInput {
     public List<Integer> getNumbers() {
         ArrayList<Integer> returnNumbers = new ArrayList<>();
         String tempStringInput = getInput();
-        while (!isValidNumberInput(tempStringInput, Constant.inputLength)) {
+        while (!validator.isValidNumberInput(tempStringInput, Constant.inputLength)) {
             Display.showWrongInputDisplay();
             tempStringInput = getInput();
         }
@@ -50,13 +30,10 @@ public class UserInput {
         return returnNumbers;
     }
 
-    public boolean isNeither(Object a, Object b, Object input) {
-        return !(input.equals(a) || input.equals(b));
-    }
-
     public boolean checkReplay() {
+        Display.showPlayAgain(REPLAY_VALUE, QUIT_VALUE);
         String userInput = getInput();
-        while (isNeither(REPLAY_VALUE, QUIT_VALUE, userInput)) {
+        while (validator.isNeither(REPLAY_VALUE, QUIT_VALUE, userInput)) {
             Display.showWrongInputDisplay();
             userInput = getInput();
         }
