+++ src/test/java/gmbs/InputValidatorTest.java
@@ -6,22 +6,22 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class UserInputTest {
-    private final UserInput userInput = new UserInput();
+class InputValidatorTest {
+    private final InputValidator validator = new InputValidator();
 
     @ParameterizedTest
     @DisplayName("입력값이 자연수인지 확인한다")
     @CsvSource(value = {"123, true", "123.4, false", "123f, false", "-123, false", "abc, false", "0, false"})
     void isNaturalNumber(String input, boolean expect) {
-        boolean actual = userInput.isNaturalNumber(input);
+        boolean actual = validator.isNaturalNumber(input);
         assertThat(actual).isEqualTo(expect);
     }
 
     @ParameterizedTest
-    @DisplayName("입력값에 중복이 있는지 확인한다")
+    @DisplayName("입력값에 중복이 없는지 확인한다")
     @CsvSource(value = {"1123, false", "asdf, true", "13a3, false", "31, true"})
     void hasNoOverlap(String input, boolean expect) {
-        boolean actual = userInput.hasNoOverlap(input);
+        boolean actual = validator.hasNoOverlap(input);
         assertThat(actual).isEqualTo(expect);
     }
 
@@ -29,7 +29,14 @@ class UserInputTest {
     @DisplayName("주어진 길이에 맞는 각 자리수가 서로 다른 자연수인지 확인한다")
     @CsvSource(value = {"111, 3, false", "1, 1, true", "123, 4, false", "123, 3, true", "0, 1, false"})
     void isValidNumberInput(String input, int length, boolean expect) {
-        boolean actual = userInput.isValidNumberInput(input, length);
+        boolean actual = validator.isValidNumberInput(input, length);
         assertThat(actual).isEqualTo(expect);
+    }
+
+    @ParameterizedTest
+    @DisplayName("input이 a와 b 둘 다 아닌지 확인한다")
+    @CsvSource(value = {"a, b, b, false", "1, 2, 3, true", "1, 2, 1, false", "b, b, a, true", "a, b, c, true"})
+    void isNeither(Object a, Object b, Object input, boolean expect) {
+        assertThat(validator.isNeither(a, b, input)).isEqualTo(expect);
     }
 }
